pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "adk-fluent"
version = "0.4.0"
description = "Fluent builder API for Google's Agent Development Kit (ADK)"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "adk-fluent contributors" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: Pydantic",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
keywords = ["adk", "google", "agents", "fluent", "builder", "llm"]

dependencies = [
    "google-adk>=1.20.0",
]

[project.optional-dependencies]
yaml = [
    "pyyaml>=6.0",
]
examples = [
    "python-dotenv>=1.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pyright>=1.1",
    "ruff>=0.9",
    "pre-commit>=3.0",
]
docs = [
    "sphinx>=7.0",
    "myst-parser>=3.0",
    "sphinx-design>=0.6",
    "furo>=2024.0",
    "sphinx-copybutton>=0.5",
]

[project.urls]
Homepage = "https://github.com/vamsiramakrishnan/adk-fluent"
Repository = "https://github.com/vamsiramakrishnan/adk-fluent"
Issues = "https://github.com/vamsiramakrishnan/adk-fluent/issues"
Documentation = "https://vamsiramakrishnan.github.io/adk-fluent/"
Changelog = "https://github.com/vamsiramakrishnan/adk-fluent/blob/master/CHANGELOG.md"

[project.entry-points."adk_extensions"]
fluent = "adk_fluent"

[tool.hatch.build.targets.wheel]
packages = ["src/adk_fluent"]

[tool.hatch.build.targets.sdist]
exclude = [
    ".venv/",
    "dist/",
    ".git/",
    ".pip-cache/",
    ".pytest_cache/",
    "__pycache__/",
    "*.egg-info/",
    "manifest.json",
]

[tool.pyright]
include = ["src"]
pythonVersion = "3.11"
typeCheckingMode = "strict"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 120
extend-exclude = ["*.pyi"]   # stubs use forward-ref types without imports

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = [
    "E501",   # line length handled by formatter
    "B905",   # zip strict — not needed for 3.11+
    "SIM108", # ternary — readability preference
]

[tool.ruff.lint.per-file-ignores]
# Generated builder modules use manifest type annotations as strings
# (from __future__ import annotations) — not importable at check time.
"src/adk_fluent/agent.py" = ["F821"]
"src/adk_fluent/workflow.py" = ["F821"]
"src/adk_fluent/tool.py" = ["F821"]
"src/adk_fluent/config.py" = ["F821"]
"src/adk_fluent/runtime.py" = ["F821"]
"src/adk_fluent/service.py" = ["F821"]
"src/adk_fluent/plugin.py" = ["F821"]
"src/adk_fluent/executor.py" = ["F821"]
"src/adk_fluent/planner.py" = ["F821"]
# __init__.py is a generated re-export hub
"src/adk_fluent/__init__.py" = ["F401", "I001"]
# Generated test scaffolds use patterns ruff dislikes
"tests/generated/*.py" = ["F811", "E711", "E712", "E731"]
# Manual tests use lambdas for quick callback stubs
"tests/**/*.py" = ["E731"]
# Examples prioritize readability over import order
"examples/**/*.py" = ["E402", "I001", "F401", "E721", "B011"]

[tool.ruff.lint.isort]
known-first-party = ["adk_fluent"]

[tool.coverage.run]
source = ["src/adk_fluent"]
omit = ["src/adk_fluent/__init__.py"]

[tool.coverage.report]
show_missing = true
skip_empty = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
